base/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
crispy-forms-materialize==0.2
Django==4.1.7
django-crispy-forms==2.0
django-materialize-css==0.0.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
idna==3.4
pytz==2022.7.1
requests==2.28.2
requests-auth==6.0.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
